feat(BridgeMaker): 입력한 길이 만큼의 다리를 생성하는 기능 추가
